+++ kiplot/__main__.py
@@ -2,8 +2,8 @@
 """KiPlot: Command-line Plotting for KiCad
 
 Usage:
-  kiplot [-b BOARD] [-c CONFIG] [-d OUT_DIR] [-s PRE] [-q | -v...] [-i]
-         [TARGET...]
+  kiplot [-b BOARD] [-e SCHEMA] [-c CONFIG] [-d OUT_DIR] [-s PRE]
+         [-q | -v...] [-i] [TARGET...]
   kiplot [-c PLOT_CONFIG] --list
   kiplot --help-list-outputs
   kiplot --help-output=HELP_OUTPUT
@@ -20,6 +20,7 @@ Options:
   -b BOARD, --board-file BOARD     The PCB .kicad-pcb board file
   -c CONFIG, --plot-config CONFIG  The plotting config file to use
   -d OUT_DIR, --out-dir OUT_DIR    The output directory [default: .]
+  -e SCHEMA, --schematic SCHEMA    The schematic file (.sch)
   --help-list-outputs              List supported outputs
   --help-output HELP_OUTPUT        Help for this particular output
   --help-outputs                   List supported outputs and details
@@ -48,12 +49,15 @@ from logging import DEBUG
 # Import log first to set the domain
 from . import log
 log.set_domain('kiplot')
-from .kiplot import (GS, generate_outputs)
+from .gs import (GS)
+from .kiplot import (generate_outputs)
 from .pre_base import (BasePreFlight)
 from .config_reader import (CfgYamlReader, print_outputs_help, print_output_help, print_preflights_help)
-from .misc import (NO_PCB_FILE, EXIT_BAD_ARGS)
+from .misc import (NO_PCB_FILE, NO_SCH_FILE, EXIT_BAD_ARGS)
 from .docopt import docopt
 from .__version__ import __version__
+
+logger = None
 
 
 def list_pre_and_outs(logger, outputs):
@@ -69,11 +73,82 @@ def list_pre_and_outs(logger, outputs):
             logger.info('- '+str(o))
 
 
+def solve_schematic(a_schematic, a_board_file):
+    schematic = a_schematic
+    if not schematic and a_board_file:
+        sch = os.path.splitext(a_board_file)[0]+'.sch'
+        if os.path.isfile(sch):
+            schematic = sch
+    if not schematic:
+        schematics = glob('*.sch')
+        if len(schematics) == 1:
+            schematic = schematics[0]
+            logger.info('Using SCH file: '+schematic)
+        elif len(schematics) > 1:
+            # Look for a schematic with a PCB
+            boards = glob('*.kicad_pcb')
+            boards_schs = [x[:-9]+'sch' for x in boards]
+            for sch in schematics:
+                if sch in boards_schs:
+                    schematic = sch
+                    break
+            if schematic is None:
+                schematic = schematics[0]
+            logger.warning('More than one SCH file found in current directory.\n'
+                           '  Using '+schematic+' if you want to use another use -e option.')
+    if schematic and not os.path.isfile(schematic):
+        logger.error("Schematic file not found: "+schematic)
+        sys.exit(NO_SCH_FILE)
+    return schematic
+
+
+def solve_config(a_plot_config):
+    plot_config = a_plot_config
+    if not plot_config:
+        plot_configs = glob('*.kiplot.yaml')
+        if len(plot_configs) == 1:
+            plot_config = plot_configs[0]
+            logger.info('Using config file: '+plot_config)
+        elif len(plot_configs) > 1:
+            plot_config = plot_configs[0]
+            logger.warning('More than one config file found in current directory.\n'
+                           '  Using '+plot_config+' if you want to use another use -c option.')
+        else:
+            logger.error('No config file found (*.kiplot.yaml), use -c to specify one.')
+            sys.exit(EXIT_BAD_ARGS)
+    if not os.path.isfile(plot_config):
+        logger.error("Plot config file not found: "+plot_config)
+        sys.exit(EXIT_BAD_ARGS)
+    return plot_config
+
+
+def solve_board_file(schematic, a_board_file):
+    board_file = a_board_file
+    if not board_file and schematic:
+        pcb = os.path.splitext(schematic)[0]+'.kicad_pcb'
+        if os.path.isfile(pcb):
+            board_file = pcb
+    if not board_file:
+        board_files = glob('*.kicad_pcb')
+        if len(board_files) == 1:
+            board_file = board_files[0]
+            logger.info('Using PCB file: '+board_file)
+        elif len(board_files) > 1:
+            board_file = board_files[0]
+            logger.warning('More than one PCB file found in current directory.\n'
+                           '  Using '+board_file+' if you want to use another use -b option.')
+    if board_file and not os.path.isfile(board_file):
+        logger.error("Board file not found: "+board_file)
+        sys.exit(NO_PCB_FILE)
+    return board_file
+
+
 def main():
     ver = 'KiPlot '+__version__+' - '+__copyright__+' - License: '+__license__
     args = docopt(__doc__, version=ver, options_first=True)
 
     # Create a logger with the specified verbosity
+    global logger
     logger = log.init(args.verbose, args.quiet)
     GS.debug_enabled = logger.getEffectiveLevel() <= DEBUG
 
@@ -91,24 +166,7 @@ def main():
         sys.exit(0)
 
     # Determine the YAML file
-    if args.plot_config is None:
-        plot_configs = glob('*.kiplot.yaml')
-        if len(plot_configs) == 1:
-            plot_config = plot_configs[0]
-            logger.info('Using config file: '+plot_config)
-        elif len(plot_configs) > 1:
-            plot_config = plot_configs[0]
-            logger.warning('More than one config file found in current directory.\n'
-                           '  Using '+plot_config+' if you want to use another use -c option.')
-        else:
-            logger.error('No config file found (*.kiplot.yaml), use -c to specify one.')
-            sys.exit(EXIT_BAD_ARGS)
-    else:
-        plot_config = args.plot_config
-    if not os.path.isfile(plot_config):
-        logger.error("Plot config file not found: "+plot_config)
-        sys.exit(EXIT_BAD_ARGS)
-
+    plot_config = solve_config(args.plot_config)
     # Read the config file
     cr = CfgYamlReader()
     outputs = None
@@ -127,29 +185,10 @@ def main():
         list_pre_and_outs(logger, outputs)
         sys.exit(0)
 
+    # Determine the SCH file
+    GS.sch_file = solve_schematic(args.schematic, args.board_file)
     # Determine the PCB file
-    if args.board_file is None:
-        board_files = glob('*.kicad_pcb')
-        if len(board_files) == 1:
-            board_file = board_files[0]
-            logger.info('Using PCB file: '+board_file)
-        elif len(board_files) > 1:
-            board_file = board_files[0]
-            logger.warning('More than one PCB file found in current directory.\n'
-                           '  Using '+board_file+' if you want to use another use -b option.')
-        else:
-            logger.error('No PCB file found (*.kicad_pcb), use -b to specify one.')
-            sys.exit(EXIT_BAD_ARGS)
-    else:
-        board_file = args.board_file
-    if not os.path.isfile(board_file):
-        logger.error("Board file not found: "+board_file)
-        sys.exit(NO_PCB_FILE)
-    GS.pcb_file = board_file
-    GS.sch_file = os.path.splitext(board_file)[0] + '.sch'
-    if not os.path.isfile(GS.sch_file):
-        logger.warning('Missing schematic file: ' + GS.sch_file)
-        GS.sch_file = None
+    GS.pcb_file = solve_board_file(GS.sch_file, args.board_file)
 
     generate_outputs(outputs, args.target, args.invert_sel, args.skip_pre)