+++ kiplot/out_pdf_pcb_print.py
@@ -2,7 +2,8 @@ import os
 from subprocess import (call)
 from .pre_base import BasePreFlight
 from .error import (KiPlotConfigurationError)
-from .kiplot import (check_script, GS)
+from .gs import (GS)
+from .kiplot import (check_script)
 from .misc import (CMD_PCBNEW_PRINT_LAYERS, URL_PCBNEW_PRINT_LAYERS, PDF_PCB_PRINT)
 from kiplot.macros import macros, document, output_class  # noqa: F401
 from . import log