+++ kiplot/pre_update_xml.py
@@ -2,7 +2,8 @@ from sys import (exit)
 from subprocess import (call)
 from kiplot.macros import macros, pre_class  # noqa: F401
 from .error import (KiPlotConfigurationError)
-from .kiplot import (GS, check_eeschema_do)
+from .gs import (GS)
+from .kiplot import (check_eeschema_do)
 from .misc import (CMD_EESCHEMA_DO, BOM_ERROR)
 from .log import (get_logger)